need to figure out a way to have the logic change if the file is changed
