+++ src/components/panels/xdc/ExperimentalTable.jsx
@@ -8,7 +8,7 @@ import * as XLSX from 'xlsx'
 import { useState } from 'react'
 
 
-export default function ExperimentalTable({onInsertFilesReady}) {
+export default function ExperimentalTable({ onInsertFilesReady, newCollectionname, newDescriptionName }) {
 
     const panelId = useContext(PanelContext)
 
@@ -18,52 +18,6 @@ export default function ExperimentalTable({onInsertFilesReady}) {
 
     const [XDCdataID] = usePanelProperty(panelId, 'results', false)
     const XDCdataFile = useFile(XDCdataID)
-
-    const [libraryName, setLibraryName] = useState(null)
-    const [description, setDescription] = useState(null)
-
-
-    //excel information
-    const readExcelFile = (eFile) => {
-        return new Promise((resolve, reject) => {
-            const reader = new FileReader()
-            reader.readAsArrayBuffer(eFile)
-            reader.onload= (event) => {resolve(event.target.result)}
-            reader.onerror = (error) =>{reject(error)}
-            })
-        }
-
-    if (experimentalFile) {
-        experimentalFile.getFile().then((realFile) => {
-            readExcelFile(realFile).then(arrayBuffer => {
-                const workbook = XLSX.read(arrayBuffer, { type: "array" });
-                const sheetName = workbook.SheetNames[0];
-                const worksheet = workbook.Sheets[sheetName];
-                const rows = XLSX.utils.sheet_to_json(worksheet, { header: 1 });
-
-                let temp_libraryName = null;
-                let temp_description = null;
-
-                for (const row of rows) {
-                    for (let i = 0; i < row.length; i++) {
-                        if (row[i] && typeof row[i] === "string") {
-                            const cell = row[i].toLowerCase();
-                            if (cell.includes("library name") || cell.includes("collection name")) {
-                                temp_libraryName = row[i+1];
-                            }
-                            if (cell.includes("description")) {
-                                temp_description = row[i+1];
-                            }
-                        }
-                        if (temp_libraryName && temp_description) {
-                            setLibraryName(temp_libraryName)
-                            setDescription(temp_description)
-                        }
-                    }
-                }
-            })
-        })
-    }
 
     return (
         <Container>
@@ -99,35 +53,19 @@ export default function ExperimentalTable({onInsertFilesReady}) {
                     </tr>
                     <tr>
                         <td><Text weight={600}> Collection Name:</Text></td>
-                        {libraryName ? 
                         <td>
                             <Group>
-                                <Text weight={600}>{libraryName}</Text>
+                                <Text weight={600}>{newCollectionname}</Text>
                             </Group>
                         </td>
-                        :
-                        <td>
-                            <Group>
-                                <Text weight={600}>{"Loading..."}</Text>
-                            </Group>
-                        </td>
-                        }
                     </tr>
                     <tr>
                         <td><Text weight={600}> Description:</Text></td>
-                        {description ?
                         <td>
                             <Group>
-                                <Text weight={600}>{description}</Text>
+                                <Text weight={600}>{newDescriptionName}</Text>
                             </Group>
                         </td>
-                        :
-                        <td>
-                            <Group>
-                                <Text weight={600}>{"Loading..."}</Text>
-                            </Group>    
-                        </td>                   
-                        }
                     </tr>
                 </tbody>
             </Table>
